requirements.txt
# Core FastAPI Stack
fastapi==0.115.12
uvicorn==0.34.2
sqlalchemy==2.0.40
pydantic==2.11.4
python-dotenv==1.1.0

# Autenticação e Segurança
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-jose[cryptography]==3.3.0

# Banco de Dados
alembic==1.16.4

# Dependências auxiliares
python-multipart==0.0.20
